minimal-requirements.txt
boto3==1.17.108
botocore==1.20.108
discord.py==2.0.0
emoji==0.4.5
praw==6.2.0
prawcore==1.0.1
PyNaCl==1.5.0